Work on setting up finder scope component to be able to find nearest item.
